refactor code, remove circular dependancies
