+++ primitives/src/lib.rs
@@ -1,10 +1,8 @@
 use ed25519_dalek::{Signature, Verifier, VerifyingKey};
-use ethers::types::H256;
+use alloy_sol_types:H256;
 use types::CircuitJustification;
 pub mod types;
 use sha2::{Digest as Sha256Digest, Sha256};
-use sp1_vectorx_script::decode_precommit;
-use sp1_vectorx_rotate_program::compute_authority_set_commitment;
 
 /// This function is useful for verifying that a Ed25519 signature is valid, it will panic if the signature is not valid
 pub fn verify_signature(pubkey_bytes: &[u8; 32], signed_message: &[u8], signature: &[u8; 64]) {
@@ -32,6 +30,54 @@ pub fn verify_simple_justification(justification: CircuitJustification, authorit
 
 }   
 
+/// Compute the new authority set hash.
+pub fn compute_authority_set_commitment(
+    num_active_authorities: usize,
+    pubkeys: Vec<[u8; 32]>,
+) -> Vec<u8> {
+    assert!(
+        num_active_authorities > 0,
+        "There must be at least one authority"
+    );
+    let mut commitment_so_far = Sha256::digest(pubkeys[0]).to_vec();
+    for pubkey in pubkeys.iter().skip(1) {
+        let mut input_to_hash = Vec::new();
+        input_to_hash.extend_from_slice(&commitment_so_far);
+        input_to_hash.extend_from_slice(pubkey);
+        commitment_so_far = Sha256::digest(&input_to_hash).to_vec();
+    }
+    commitment_so_far
+}
+
+pub fn decode_precommit(precommit: Vec<u8>) -> (H256, u32, u64, u64) {
+    // The first byte should be a 1.
+    assert_eq!(precommit[0], 1);
+
+    // The next 32 bytes are the block hash.
+    let block_hash = &precommit[1..33];
+
+    // The next 4 bytes are the block number.
+    let block_number = &precommit[33..37];
+    // Convert the block number to a u32.
+    let block_number = u32::from_le_bytes(block_number.try_into().unwrap());
+
+    // The next 8 bytes are the justification round.
+    let round = &precommit[37..45];
+    // Convert the round to a u64.
+    let round = u64::from_le_bytes(round.try_into().unwrap());
+
+    // The next 8 bytes are the authority set id.
+    let authority_set_id = &precommit[45..53];
+    // Convert the authority set id to a u64.
+    let authority_set_id = u64::from_le_bytes(authority_set_id.try_into().unwrap());
+
+    (
+        H256::from_slice(block_hash),
+        block_number,
+        round,
+        authority_set_id,
+    )
+}
 
 /// Decode a SCALE-encoded compact int.
 pub fn decode_scale_compact_int(bytes: &[u8]) -> (u64, usize) {
